Split classes into files
