+++ osu.Framework/Graphics/Batches/VertexBatch.cs
@@ -181,68 +181,4 @@ namespace osu.Framework.Graphics.Batches
             return ref vertices;
         }
     }
-
-    public interface IVertexGroup<in T>
-    {
-        void Add(T vertex);
-    }
-
-    public readonly struct DelegatingVertexGroup<T> : IVertexGroup<T>
-        where T : struct, IEquatable<T>, IVertex
-    {
-        private readonly Action<T> action;
-
-        public DelegatingVertexGroup(Action<T> action)
-        {
-            this.action = action;
-        }
-
-        public void Add(T vertex) => action(vertex);
-    }
-
-    public struct VertexGroup<T> : IVertexGroup<T>, IDisposable
-        where T : struct, IEquatable<T>, IVertex
-    {
-        internal readonly VertexBatch<T> Batch;
-        internal readonly long InvalidationID;
-        internal readonly int StartIndex;
-        internal readonly float DrawDepth;
-
-        internal ulong FrameIndex;
-        internal bool DrawRequired;
-        internal int Count;
-
-        public VertexGroup(VertexBatch<T> batch, long invalidationID, int startIndex, float drawDepth)
-        {
-            Batch = batch;
-            InvalidationID = invalidationID;
-            StartIndex = startIndex;
-            DrawDepth = drawDepth;
-
-            FrameIndex = 0;
-            DrawRequired = false;
-            Count = 0;
-        }
-
-        public void Add(T vertex)
-        {
-            if (DrawRequired)
-                Batch.AddVertex(vertex);
-            else
-            {
-#if VBO_CONSISTENCY_CHECKS
-                if (!Batch.GetCurrentVertex().Equals(vertex))
-                    throw new InvalidOperationException("Vertex draw was skipped, but the contained vertex differs.");
-#endif
-
-                Batch.Advance();
-            }
-
-            Count++;
-        }
-
-        public void Dispose()
-        {
-        }
-    }
 }
